refactor settings maintenance modules
