+++ src/renderer/src/components/UpdateCard.tsx
@@ -1,94 +1,31 @@
-/* eslint-disable max-lines -- Why: keeps the updater state machine and its presentation variants in one file. */
 import { useCallback, useEffect, useRef, useState } from 'react'
+import { X } from 'lucide-react'
+import type { ChangelogData } from '../../../shared/update-status-types'
 import { usePrefersReducedMotion } from '@/hooks/usePrefersReducedMotion'
 import { useAppStore } from '../store'
 import { Card } from './ui/card'
 import { Button } from './ui/button'
-import { Progress } from './ui/progress'
-import { AlertCircle, Check, Loader2, Minus, X } from 'lucide-react'
-import type { ChangelogData } from '../../../shared/update-status-types'
-import { UpdateErrorCardContent, type UpdateErrorCardModel } from './UpdateErrorCardContent'
-import { LinuxPackageInstallRecoveryCard } from './LinuxPackageInstallRecoveryCard'
-import {
-  isWindowsSignatureCheckUnavailableFailure,
-  isWindowsSignatureMismatchFailure
-} from '../../../shared/updater-windows-signature-check'
-import { getReleaseNotesUrlForVersion } from '../../../shared/release-channel'
 import { translate } from '@/i18n/i18n'
+import { UpdateCardStateContent } from './maintenance/update-card/UpdateCardStateContent'
+import { buildUpdateCardErrorModel } from './maintenance/update-card/update-card-error-model'
+import {
+  getUpdateCardAriaLabel,
+  isHttp2ProtocolError,
+  isUpdateCardVisible
+} from './maintenance/update-card/update-card-visibility'
 
-// ── Helpers ──────────────────────────────────────────────────────────
+export { isHttp2ProtocolError }
 
-function isAnimatedGif(url: string | undefined): boolean {
-  return typeof url === 'string' && url.toLowerCase().endsWith('.gif')
-}
-
-export function isHttp2ProtocolError(message: string): boolean {
-  const normalized = message.toLowerCase()
-  return (
-    normalized.includes('err_http2_protocol_error') ||
-    normalized.includes('http2_protocol_error') ||
-    (normalized.includes('http/2') && normalized.includes('protocol'))
-  )
-}
-
-// ── Compact card (transient check feedback) ─────────────────────────
-
-function CompactCardContent({
-  icon,
-  text,
-  onClose,
-  action
-}: {
-  icon: 'spinner' | 'check' | 'error'
-  text: string
-  onClose?: () => void
-  action?: { label: string; url: string }
-}) {
-  return (
-    <div className="flex items-center gap-3 p-3">
-      <div className="shrink-0 text-muted-foreground">
-        {icon === 'spinner' && <Loader2 className="size-4 animate-spin" />}
-        {icon === 'check' && <Check className="size-4" />}
-        {icon === 'error' && <AlertCircle className="size-4" />}
-      </div>
-      <div className="flex-1 min-w-0">
-        <p className="text-sm truncate">{text}</p>
-        {action && (
-          <button
-            className="text-xs text-muted-foreground underline hover:text-foreground mt-0.5"
-            onClick={() => void window.api.shell.openUrl(action.url)}
-          >
-            {action.label}
-          </button>
-        )}
-      </div>
-      {onClose && (
-        <Button
-          variant="ghost"
-          size="icon"
-          className="size-7 shrink-0"
-          onClick={onClose}
-          aria-label={translate('auto.components.UpdateCard.a726967bd3', 'Dismiss')}
-        >
-          <X className="size-3.5" />
-        </Button>
-      )}
-    </div>
-  )
-}
-
-// ── Main component ──────────────────────────────────────────────────
-
-export function UpdateCard() {
-  const status = useAppStore((s) => s.updateStatus)
-  const storeChangelog = useAppStore((s) => s.updateChangelog)
-  const updateUserInitiatedCycle = useAppStore((s) => s.updateUserInitiatedCycle)
-  const dismissedVersion = useAppStore((s) => s.dismissedUpdateVersion)
-  const dismissUpdate = useAppStore((s) => s.dismissUpdate)
-  const collapsed = useAppStore((s) => s.updateCardCollapsed)
-  const setCollapsed = useAppStore((s) => s.setUpdateCardCollapsed)
-  const reassuranceSeen = useAppStore((s) => s.updateReassuranceSeen)
-  const markReassuranceSeen = useAppStore((s) => s.markUpdateReassuranceSeen)
+export function UpdateCard(): React.JSX.Element | null {
+  const status = useAppStore((state) => state.updateStatus)
+  const changelog: ChangelogData | null = useAppStore((state) => state.updateChangelog)
+  const updateUserInitiatedCycle = useAppStore((state) => state.updateUserInitiatedCycle)
+  const dismissedVersion = useAppStore((state) => state.dismissedUpdateVersion)
+  const dismissUpdate = useAppStore((state) => state.dismissUpdate)
+  const collapsed = useAppStore((state) => state.updateCardCollapsed)
+  const setCollapsed = useAppStore((state) => state.setUpdateCardCollapsed)
+  const reassuranceSeen = useAppStore((state) => state.updateReassuranceSeen)
+  const markReassuranceSeen = useAppStore((state) => state.markUpdateReassuranceSeen)
   const hasStartedDownload = useRef(false)
   const dismissAnimationTimerRef = useRef<number | null>(null)
   const collapseAnimationTimerRef = useRef<number | null>(null)
@@ -97,16 +34,10 @@ export function UpdateCard() {
   const [installError, setInstallError] = useState<string | null>(null)
   const [compatibilityRelaunching, setCompatibilityRelaunching] = useState(false)
   const [compatibilitySetupError, setCompatibilitySetupError] = useState<string | null>(null)
-  // Why: the dismiss gate keeps error cards visible, so a separate local flag tracks the error card's own X close.
   const [errorDismissed, setErrorDismissed] = useState(false)
-  // Why: local flag (not store) for the transient "up to date" auto-dismiss — no other component needs it.
   const [autoDismissed, setAutoDismissed] = useState(false)
-  // Tracks card exit so the fade-out animation plays before unmount.
   const [exiting, setExiting] = useState(false)
-  const changelog: ChangelogData | null = storeChangelog
   const isLocalBuild = status.source === 'local'
-
-  // Why: the 'error' variant carries no version, but the card needs it for the fallback URL and dismiss; cache from states that have it.
   const versionRef = useRef<string | null>(null)
   if ('version' in status && status.version) {
     versionRef.current = status.version
@@ -115,11 +46,8 @@ export function UpdateCard() {
     status.state === 'idle' ||
     status.state === 'not-available'
   ) {
-    // Why: clear the cached version so a later check failure can't link/dismiss against an unrelated older release.
     versionRef.current = null
   }
-
-  // Why: reset component-local state on a new version so stale flags (media load, hasStartedDownload) don't leak forward.
   const prevVersionRef = useRef<string | null>(null)
   if (status.state === 'available' && status.version !== prevVersionRef.current) {
     prevVersionRef.current = status.version
@@ -128,8 +56,6 @@ export function UpdateCard() {
     setMediaLoaded(false)
     setInstallError(null)
   }
-
-  // Why: reset per-cycle flags when a new status arrives so the card shows again next check cycle.
   const prevStateRef = useRef(status.state)
   if (status.state !== prevStateRef.current) {
     prevStateRef.current = status.state
@@ -146,18 +72,13 @@ export function UpdateCard() {
 
   const shouldAutoDismissLatest =
     status.state === 'not-available' && 'userInitiated' in status && Boolean(status.userInitiated)
-
-  // Auto-dismiss "You're on the latest version" after 3s; timer resets if status changes first.
   useEffect(() => {
     if (!shouldAutoDismissLatest) {
       return
     }
-    const timer = setTimeout(() => setAutoDismissed(true), 3000)
-    return () => clearTimeout(timer)
+    const timer = window.setTimeout(() => setAutoDismissed(true), 3000)
+    return () => window.clearTimeout(timer)
   }, [shouldAutoDismissLatest])
-
-  // Why: quitAndInstall must run in an effect, not render — StrictMode's double render would fire it twice.
-  // Gated on hasStartedDownload so Settings-initiated downloads don't auto-restart (user expects "Restart" there).
   useEffect(() => {
     if (status.state === 'downloaded' && hasStartedDownload.current) {
       void window.api.updater.quitAndInstall().catch((error) => {
@@ -166,9 +87,7 @@ export function UpdateCard() {
     }
   }, [status.state])
 
-  // ── Prefers-reduced-motion ──────────────────────────────────────────
   const prefersReducedMotion = usePrefersReducedMotion()
-
   const clearAnimationTimers = useCallback(() => {
     if (dismissAnimationTimerRef.current !== null) {
       window.clearTimeout(dismissAnimationTimerRef.current)
@@ -179,81 +98,38 @@ export function UpdateCard() {
       collapseAnimationTimerRef.current = null
     }
   }, [])
-
   const cardRootRef = useCallback(
     (node: HTMLDivElement | null) => {
-      if (node !== null) {
-        return
+      if (node === null) {
+        clearAnimationTimers()
       }
-      // Why: cancel exit timers when the card surface unmounts so stale callbacks don't fire.
-      clearAnimationTimers()
     },
     [clearAnimationTimers]
   )
-
-  // ── Visibility gates ──────────────────────────────────────────────
-
-  const isUserInitiated = 'userInitiated' in status && status.userInitiated
   const cachedVersion = versionRef.current
-  const shouldShowDetailedErrorCard =
-    status.state === 'error' && (hasStartedDownload.current || cachedVersion !== null)
-
-  // Compact transient states: only show for user-initiated checks.
-  if (status.state === 'checking' && !isUserInitiated) {
-    return null
-  }
-  if (status.state === 'not-available' && !isUserInitiated) {
-    return null
-  }
-  if (status.state === 'not-available' && autoDismissed) {
-    return null
-  }
-
-  // Background states that never show the card.
-  if (status.state === 'idle') {
-    return null
-  }
-
-  // Error: show for user-initiated failures or failures tied to a cached version; background failures stay silent.
-  if (status.state === 'error' && !shouldShowDetailedErrorCard && !isUserInitiated) {
-    return null
-  }
-
-  // Why: the dismiss gate below keeps error cards visible, so an explicit X on the error card needs this gate to hide it.
-  if (status.state === 'error' && errorDismissed) {
-    return null
-  }
-
-  // Dismiss gate: hide previously-dismissed versions for passive states, keep in-progress/error visible, and bypass for user-initiated checks.
-  if (versionRef.current && dismissedVersion === versionRef.current && !updateUserInitiatedCycle) {
-    if (status.state !== 'downloading' && status.state !== 'error') {
-      return null
-    }
-  }
-
   if (
-    collapsed &&
-    (status.state === 'downloading' || status.state === 'downloaded' || status.state === 'error')
+    !isUpdateCardVisible({
+      status,
+      dismissedVersion,
+      cachedVersion,
+      hasStartedDownload: hasStartedDownload.current,
+      updateUserInitiatedCycle,
+      autoDismissed,
+      errorDismissed,
+      collapsed
+    })
   ) {
     return null
   }
 
-  // ── Shared helpers ────────────────────────────────────────────────
-
-  const isRichMode = changelog?.release != null
-
-  const handleUpdate = () => {
+  const handleUpdate = (): void => {
     hasStartedDownload.current = true
-    // Why: clicking Update implies the user isn't worried about interruption, so retire the reassurance tip.
     if (!reassuranceSeen) {
       markReassuranceSeen()
     }
     void window.api.updater.download()
   }
-
-  // Why: the 'error' variant has no version field, so dismiss needs an explicit version override.
-  const handleClose = () => {
-    // Why: dismissUpdate clears the store manual-check bypass so the dismiss gate re-engages after closing.
+  const handleClose = (): void => {
     if (status.state === 'error') {
       setErrorDismissed(true)
       if (cachedVersion) {
@@ -263,16 +139,12 @@ export function UpdateCard() {
     }
     dismissUpdate()
   }
-
-  const handleInstallRetry = () => {
+  const handleInstallRetry = (): void => {
     void window.api.updater.quitAndInstall().catch((error) => {
       setInstallError(String((error as Error)?.message ?? error))
     })
   }
-
-  const handleEnableHttp1Compatibility = () => {
-    // Why: the shared error card marks a pending action aria-disabled rather than disabled, so the
-    // second click of a double-click now reaches this handler and would relaunch twice.
+  const handleEnableHttp1Compatibility = (): void => {
     if (compatibilityRelaunching) {
       return
     }
@@ -288,132 +160,25 @@ export function UpdateCard() {
         setCompatibilityRelaunching(false)
       })
   }
-
-  // Why: order matters — the wrong-publisher security-stop must beat the "check couldn't run" case so integrity failures aren't softened to "try again".
-  const isHttp2UpdateError = status.state === 'error' && isHttp2ProtocolError(status.message)
-  const isSignatureMismatchError =
-    status.state === 'error' && isWindowsSignatureMismatchFailure(status.message)
-  const isSignatureCheckBlockedError =
-    status.state === 'error' && isWindowsSignatureCheckUnavailableFailure(status.message)
-  // Carries the diagnostic alongside the recovery so the render branch needs no second state check.
+  const errorCard = buildUpdateCardErrorModel({
+    status,
+    isLocalBuild,
+    cachedVersion,
+    installError,
+    compatibilityRelaunching,
+    compatibilitySetupError,
+    onChooseLocalBuild: () => void window.api.updater.check({ localBuild: true }),
+    onEnableHttp1Compatibility: handleEnableHttp1Compatibility,
+    onRetryDownload: handleUpdate,
+    onRecheck: () => void window.api.updater.check({ includePrerelease: false }),
+    onInstallRetry: handleInstallRetry
+  })
   const linuxPackageRecovery =
     status.state === 'error' && status.recovery?.kind === 'linux-package-install'
       ? { recovery: status.recovery, diagnostic: status.message }
       : null
-  const errorCard: UpdateErrorCardModel | null =
-    status.state === 'error'
-      ? isLocalBuild
-        ? {
-            title: cachedVersion
-              ? translate('auto.components.UpdateCard.8cf17b10af', 'Local Build Error')
-              : translate('auto.components.UpdateCard.a4650b0dc4', 'Could Not Use Local Build'),
-            summary: cachedVersion
-              ? translate(
-                  'auto.components.UpdateCard.b1e390250d',
-                  'Could not complete the local build switch.'
-                )
-              : translate(
-                  'auto.components.UpdateCard.d29740d175',
-                  'The selected build could not be used.'
-                ),
-            detail: status.message,
-            primaryAction: {
-              label: translate('auto.components.UpdateCard.37d45c9ec1', 'Choose Another Build'),
-              onClick: () => {
-                void window.api.updater.check({ localBuild: true })
-              }
-            }
-          }
-        : isHttp2UpdateError
-          ? {
-              variant: 'http1Compatibility',
-              title: translate('auto.components.UpdateCard.1339b82cee', 'HTTP/2 Download Blocked'),
-              summary: 'Orca can retry through HTTP/1.1 compatibility mode.',
-              explainer: translate(
-                'auto.components.UpdateCard.90559b14e3',
-                'This turns on a process-wide Electron networking switch after restart. Use it for corporate VPNs or proxies that reject HTTP/2 update downloads.'
-              ),
-              detail: compatibilitySetupError ?? status.message,
-              releaseUrl: getReleaseNotesUrlForVersion(cachedVersion),
-              primaryAction: {
-                label: translate('auto.components.UpdateCard.933c6fdf5b', 'Enable & Restart'),
-                pendingLabel: 'Restarting...',
-                isPending: compatibilityRelaunching,
-                onClick: handleEnableHttp1Compatibility
-              }
-            }
-          : isSignatureMismatchError
-            ? {
-                // Security stop: installer signed by the wrong publisher — no retry, only a verified-download path.
-                variant: 'security',
-                title: translate(
-                  'auto.components.UpdateCard.5b309b19f3',
-                  "Update Wasn't Installed"
-                ),
-                summary: translate(
-                  'auto.components.UpdateCard.092f09fc14',
-                  "The installer's publisher doesn't match Orca, so we stopped the update. Don't install this download; check official releases for a corrected version."
-                ),
-                detail: status.message,
-                // Why: linking the rejected version would let users bypass the publisher check by re-running it.
-                releaseUrl: getReleaseNotesUrlForVersion(null),
-                manualLabel: translate(
-                  'auto.components.UpdateCard.c9ff9b9ec2',
-                  'Check official releases'
-                )
-              }
-            : isSignatureCheckBlockedError
-              ? {
-                  title: translate(
-                    'auto.components.UpdateCard.e944c2de43',
-                    'Update Verification Blocked'
-                  ),
-                  summary: translate(
-                    'auto.components.UpdateCard.a05992a26b',
-                    "The signature check couldn't run — usually because antivirus software blocked it. Retry the download, or get the installer from our official releases."
-                  ),
-                  detail: status.message,
-                  releaseUrl: getReleaseNotesUrlForVersion(cachedVersion),
-                  primaryAction: {
-                    label: translate('auto.components.UpdateCard.48565a32bc', 'Retry Download'),
-                    onClick: handleUpdate
-                  }
-                }
-              : {
-                  // Why: title is scoped to the failed operation so check-time (GitHub-side) failures don't read as an Orca bug.
-                  title: cachedVersion ? 'Update Error' : 'Update Check Failed',
-                  summary: cachedVersion
-                    ? 'Could not complete the update.'
-                    : 'Could not check for updates.',
-                  detail: status.message,
-                  releaseUrl: getReleaseNotesUrlForVersion(cachedVersion),
-                  // Why: check-time failures are often transient, so offer a Re-check instead of forcing manual download.
-                  primaryAction: cachedVersion
-                    ? {
-                        label: translate('auto.components.UpdateCard.48565a32bc', 'Retry Download'),
-                        onClick: handleUpdate
-                      }
-                    : {
-                        label: translate('auto.components.UpdateCard.6b0085010d', 'Re-check'),
-                        onClick: () => {
-                          void window.api.updater.check({ includePrerelease: false })
-                        }
-                      }
-                }
-      : installError
-        ? {
-            title: translate('auto.components.UpdateCard.4cf109845a', 'Update Error'),
-            summary: 'Could not restart to install the update.',
-            detail: installError,
-            releaseUrl: getReleaseNotesUrlForVersion(cachedVersion),
-            primaryAction: {
-              label: translate('auto.components.UpdateCard.2c2d3e03ca', 'Try Again'),
-              onClick: handleInstallRetry
-            }
-          }
-        : null
 
-  const handleDismissWithAnimation = () => {
+  const handleDismissWithAnimation = (): void => {
     if (prefersReducedMotion) {
       handleClose()
       return
@@ -427,9 +192,7 @@ export function UpdateCard() {
       handleClose()
     }, 150)
   }
-
-  // Why: dismissing an active download would orphan it, so long-running phases minimize to the status bar.
-  const handleCollapseWithAnimation = () => {
+  const handleCollapseWithAnimation = (): void => {
     if (prefersReducedMotion) {
       setCollapsed(true)
       return
@@ -444,12 +207,11 @@ export function UpdateCard() {
       setExiting(false)
     }, 150)
   }
-
-  const handleKeyDown = (e: React.KeyboardEvent) => {
-    if (e.key !== 'Escape') {
+  const handleKeyDown = (event: React.KeyboardEvent): void => {
+    if (event.key !== 'Escape') {
       return
     }
-    e.preventDefault()
+    event.preventDefault()
     if (
       status.state === 'downloading' ||
       status.state === 'downloaded' ||
@@ -461,157 +223,38 @@ export function UpdateCard() {
     }
   }
 
-  // ── Dynamic aria-label ────────────────────────────────────────────
-
-  const ariaLabel =
-    status.state === 'checking'
-      ? 'Checking for updates'
-      : status.state === 'not-available'
-        ? "You're on the latest version"
-        : status.state === 'available'
-          ? 'Update available'
-          : status.state === 'downloading'
-            ? 'Downloading update'
-            : status.state === 'downloaded'
-              ? 'Update ready to install'
-              : status.state === 'error'
-                ? 'Update error'
-                : 'Update status'
-
-  // ── Card wrapper ──────────────────────────────────────────────────
+  const cardContent = (
+    <UpdateCardStateContent
+      status={status}
+      changelog={changelog}
+      errorCard={errorCard}
+      linuxPackageRecovery={linuxPackageRecovery}
+      isLocalBuild={isLocalBuild}
+      cachedVersion={cachedVersion}
+      hasStartedDownload={hasStartedDownload.current}
+      prefersReducedMotion={prefersReducedMotion}
+      mediaFailed={mediaFailed}
+      mediaLoaded={mediaLoaded}
+      onMediaError={() => setMediaFailed(true)}
+      onMediaLoad={() => setMediaLoaded(true)}
+      onUpdate={handleUpdate}
+      onInstallRetry={handleInstallRetry}
+      onDismiss={handleDismissWithAnimation}
+      onCollapse={handleCollapseWithAnimation}
+    />
+  )
 
   const animationClass = prefersReducedMotion
     ? ''
     : exiting
       ? 'animate-update-card-exit'
       : 'animate-update-card-enter'
-
-  const cardContent = (() => {
-    // ── Compact transient states (user-initiated check feedback) ──────
-
-    if (status.state === 'checking') {
-      return (
-        <CompactCardContent
-          icon="spinner"
-          text={translate('auto.components.UpdateCard.ba5ffc949c', 'Checking for updates...')}
-        />
-      )
-    }
-
-    if (status.state === 'not-available') {
-      return (
-        <CompactCardContent
-          icon="check"
-          text={translate('auto.components.UpdateCard.ea2a41adbe', "You're on the latest version.")}
-        />
-      )
-    }
-
-    // ── Error states ─────────────────────────────────────────────────
-
-    // Why: the package-recovery card owns its own async validation state, so it branches before the generic models.
-    if (linuxPackageRecovery) {
-      return (
-        <LinuxPackageInstallRecoveryCard
-          recovery={linuxPackageRecovery.recovery}
-          diagnostic={linuxPackageRecovery.diagnostic}
-          releaseUrl={isLocalBuild ? undefined : getReleaseNotesUrlForVersion(cachedVersion)}
-          onClose={handleCollapseWithAnimation}
-        />
-      )
-    }
-
-    if (errorCard) {
-      return <UpdateErrorCardContent {...errorCard} onClose={handleCollapseWithAnimation} />
-    }
-
-    // ── Downloaded state ─────────────────────────────────────────────
-
-    if (status.state === 'downloaded') {
-      if (hasStartedDownload.current) {
-        return (
-          <div className="p-4">
-            <p className="text-sm">
-              {translate('auto.components.UpdateCard.09a55c39b5', 'Installing...')}
-            </p>
-          </div>
-        )
-      }
-      // Settings-initiated download — show "Ready to install"
-      return (
-        <ReadyToInstallContent
-          version={status.version}
-          onRestart={handleInstallRetry}
-          onClose={handleCollapseWithAnimation}
-        />
-      )
-    }
-
-    // ── Downloading state ────────────────────────────────────────────
-
-    if (status.state === 'downloading') {
-      return (
-        <DownloadingContent
-          version={status.version}
-          percent={status.percent}
-          changelog={changelog}
-          prefersReducedMotion={prefersReducedMotion}
-          mediaFailed={mediaFailed}
-          mediaLoaded={mediaLoaded}
-          onMediaError={() => setMediaFailed(true)}
-          onMediaLoad={() => setMediaLoaded(true)}
-          onCollapse={handleCollapseWithAnimation}
-          showReleaseNotes={!isLocalBuild}
-        />
-      )
-    }
-
-    // ── Available state ──────────────────────────────────────────────
-
-    if (status.state !== 'available') {
-      return null
-    }
-
-    const releaseUrl = isLocalBuild
-      ? undefined
-      : (('releaseUrl' in status ? status.releaseUrl : undefined) ??
-        getReleaseNotesUrlForVersion(status.version))
-
-    if (isRichMode && changelog) {
-      return (
-        <RichCardContent
-          release={changelog.release}
-          releasesBehind={changelog.releasesBehind}
-          prefersReducedMotion={prefersReducedMotion}
-          mediaFailed={mediaFailed}
-          mediaLoaded={mediaLoaded}
-          onMediaError={() => setMediaFailed(true)}
-          onMediaLoad={() => setMediaLoaded(true)}
-          onUpdate={handleUpdate}
-          onClose={handleDismissWithAnimation}
-        />
-      )
-    }
-
-    return (
-      <SimpleCardContent
-        version={status.version}
-        releaseUrl={releaseUrl}
-        onUpdate={handleUpdate}
-        onClose={handleDismissWithAnimation}
-      />
-    )
-  })()
-
-  // One-time reassurance tip that updating won't kill running terminals; persisted once seen.
   const showReassurance =
     !reassuranceSeen && (status.state === 'available' || status.state === 'downloading')
-
   return (
     <div
       ref={cardRootRef}
-      className="fixed bottom-10 right-4 z-40 w-[360px] max-w-[calc(100vw-32px)] flex flex-col gap-2
-      max-[480px]:left-4 max-[480px]:right-4 max-[480px]:w-auto"
+      className="fixed bottom-10 right-4 z-40 w-[360px] max-w-[calc(100vw-32px)] flex flex-col gap-2 max-[480px]:left-4 max-[480px]:right-4 max-[480px]:w-auto"
     >
       {showReassurance && (
         <Card className={`py-0 gap-0 ${animationClass}`}>
@@ -638,7 +281,7 @@ export function UpdateCard() {
       )}
       <Card
         role="complementary"
-        aria-label={ariaLabel}
+        aria-label={getUpdateCardAriaLabel(status)}
         aria-live="polite"
         tabIndex={-1}
         onKeyDown={handleKeyDown}
@@ -646,310 +289,6 @@ export function UpdateCard() {
       >
         {cardContent}
       </Card>
-    </div>
-  )
-}
-
-// ── Rich card content ────────────────────────────────────────────────
-
-function RichCardContent({
-  release,
-  releasesBehind,
-  prefersReducedMotion,
-  mediaFailed,
-  mediaLoaded,
-  onMediaError,
-  onMediaLoad,
-  onUpdate,
-  onClose
-}: {
-  release: NonNullable<ChangelogData['release']>
-  releasesBehind: number | null
-  prefersReducedMotion: boolean
-  mediaFailed: boolean
-  mediaLoaded: boolean
-  onMediaError: () => void
-  onMediaLoad: () => void
-  onUpdate: () => void
-  onClose: () => void
-}) {
-  const showMedia =
-    release.mediaUrl &&
-    !mediaFailed &&
-    // Why: GIFs can't be reliably paused cross-browser, so hide them entirely under reduced-motion.
-    !(prefersReducedMotion && isAnimatedGif(release.mediaUrl))
-
-  return (
-    <div className="flex flex-col gap-3 p-4">
-      <div className="flex items-start justify-between gap-2">
-        <h3 className="text-sm font-semibold">
-          {translate('auto.components.UpdateCard.f58b5c57a6', 'New:')} {release.title}
-        </h3>
-        <Button
-          variant="ghost"
-          size="icon"
-          className="size-7 shrink-0 min-w-[44px] min-h-[44px] -m-2"
-          onClick={onClose}
-          aria-label={translate('auto.components.UpdateCard.318d3b4bc7', 'Dismiss update')}
-        >
-          <X className="size-3.5" />
-        </Button>
-      </div>
-
-      {showMedia && (
-        <div className="relative overflow-hidden rounded-md">
-          {!mediaLoaded && (
-            // Shimmer placeholder while image loads
-            <div
-              className="w-full bg-muted/50 animate-pulse rounded-md"
-              style={{ aspectRatio: '16/9' }}
-            />
-          )}
-          <img
-            src={release.mediaUrl}
-            alt=""
-            className={`w-full rounded-md ${mediaLoaded ? '' : 'absolute inset-0'}`}
-            style={!mediaLoaded ? { visibility: 'hidden' } : undefined}
-            onError={onMediaError}
-            onLoad={onMediaLoad}
-          />
-        </div>
-      )}
-
-      <p className="text-sm text-muted-foreground">
-        {release.description}
-        {releasesBehind !== null && releasesBehind > 1 && (
-          <>
-            {' '}
-            <button
-              className="text-xs text-muted-foreground/70 underline hover:text-foreground inline"
-              onClick={() => void window.api.shell.openUrl(release.releaseNotesUrl)}
-            >
-              +{releasesBehind - 1}{' '}
-              {translate('auto.components.UpdateCard.ccd8b0a793', 'more since your last update')}
-            </button>
-          </>
-        )}
-      </p>
-
-      <button
-        className="text-xs text-muted-foreground underline hover:text-foreground self-start"
-        onClick={() => void window.api.shell.openUrl(release.releaseNotesUrl)}
-      >
-        {translate('auto.components.UpdateCard.aad383aecc', 'Read the full release notes')}
-      </button>
-
-      <Button variant="default" size="sm" onClick={onUpdate} className="w-full cursor-pointer">
-        {translate('auto.components.UpdateCard.ec8fe71cfc', 'Update')}
-      </Button>
-    </div>
-  )
-}
-
-// ── Simple card content ──────────────────────────────────────────────
-
-function SimpleCardContent({
-  version,
-  releaseUrl,
-  onUpdate,
-  onClose
-}: {
-  version: string
-  releaseUrl?: string
-  onUpdate: () => void
-  onClose: () => void
-}) {
-  return (
-    <div className="flex flex-col gap-2.5 p-3.5">
-      <div className="flex items-start justify-between gap-2">
-        <h3 className="text-sm font-semibold">
-          {translate('auto.components.UpdateCard.9abc59f814', 'Update Available')}
-        </h3>
-        <Button
-          variant="ghost"
-          size="icon"
-          className="size-7 shrink-0 min-w-[44px] min-h-[44px] -m-2"
-          onClick={onClose}
-          aria-label={translate('auto.components.UpdateCard.318d3b4bc7', 'Dismiss update')}
-        >
-          <X className="size-3.5" />
-        </Button>
-      </div>
-
-      <p className="text-sm text-muted-foreground">
-        {translate('auto.components.UpdateCard.05ad78a6d1', 'Orca v{{value0}} is ready.', {
-          value0: version
-        })}
-      </p>
-
-      <p className="text-xs leading-relaxed text-muted-foreground">
-        {translate('auto.components.UpdateCard.fdd4a364fa', "Sessions won't be interrupted.")}
-      </p>
-
-      {releaseUrl && (
-        <button
-          className="text-xs text-muted-foreground underline underline-offset-2 hover:text-foreground self-start"
-          onClick={() => void window.api.shell.openUrl(releaseUrl)}
-        >
-          {translate('auto.components.UpdateCard.44324ef542', 'Release notes')}
-        </button>
-      )}
-
-      <Button
-        variant="default"
-        size="sm"
-        onClick={onUpdate}
-        className="mt-0.5 w-full cursor-pointer"
-      >
-        {translate('auto.components.UpdateCard.ec8fe71cfc', 'Update')}
-      </Button>
-    </div>
-  )
-}
-
-// ── Downloading content ──────────────────────────────────────────────
-
-function DownloadingContent({
-  version,
-  percent,
-  changelog,
-  prefersReducedMotion,
-  mediaFailed,
-  mediaLoaded,
-  onMediaError,
-  onMediaLoad,
-  onCollapse,
-  showReleaseNotes
-}: {
-  version: string
-  percent: number
-  changelog: ChangelogData | null
-  prefersReducedMotion: boolean
-  mediaFailed: boolean
-  mediaLoaded: boolean
-  onMediaError: () => void
-  onMediaLoad: () => void
-  onCollapse: () => void
-  showReleaseNotes: boolean
-}) {
-  const release = changelog?.release
-  const showMedia =
-    release?.mediaUrl && !mediaFailed && !(prefersReducedMotion && isAnimatedGif(release.mediaUrl))
-
-  return (
-    <div className="flex flex-col gap-3 p-4">
-      <div className="flex items-start justify-between gap-2">
-        {release ? (
-          <h3 className="text-sm font-semibold">
-            {translate('auto.components.UpdateCard.f58b5c57a6', 'New:')} {release.title}
-          </h3>
-        ) : (
-          <h3 className="text-sm font-semibold">
-            {translate('auto.components.UpdateCard.558842597d', 'Downloading Update')}
-          </h3>
-        )}
-        <Button
-          variant="ghost"
-          size="icon"
-          className="size-7 shrink-0 min-w-[44px] min-h-[44px] -m-2"
-          onClick={onCollapse}
-          aria-label={translate('auto.components.UpdateCard.8acbdd3961', 'Minimize to status bar')}
-        >
-          <Minus className="size-3.5" />
-        </Button>
-      </div>
-
-      {showMedia && release?.mediaUrl && (
-        <div className="relative overflow-hidden rounded-md">
-          {!mediaLoaded && (
-            <div
-              className="w-full bg-muted/50 animate-pulse rounded-md"
-              style={{ aspectRatio: '16/9' }}
-            />
-          )}
-          <img
-            src={release.mediaUrl}
-            alt=""
-            className={`w-full rounded-md ${mediaLoaded ? '' : 'absolute inset-0'}`}
-            style={!mediaLoaded ? { visibility: 'hidden' } : undefined}
-            onError={onMediaError}
-            onLoad={onMediaLoad}
-          />
-        </div>
-      )}
-
-      <p className="text-sm text-muted-foreground">
-        {release
-          ? release.description
-          : translate('auto.components.UpdateCard.93794ea932', 'Orca v{{value0}} is downloading.', {
-              value0: version
-            })}
-      </p>
-
-      {showReleaseNotes && (
-        <button
-          className="text-xs text-muted-foreground underline hover:text-foreground self-start"
-          onClick={() =>
-            void window.api.shell.openUrl(
-              release ? release.releaseNotesUrl : getReleaseNotesUrlForVersion(version)
-            )
-          }
-        >
-          {release
-            ? translate('auto.components.UpdateCard.aad383aecc', 'Read the full release notes')
-            : translate('auto.components.UpdateCard.44324ef542', 'Release notes')}
-        </button>
-      )}
-
-      <div className="flex flex-col gap-2 mt-1">
-        <Progress value={percent} className="h-1.5" />
-        <p className="text-xs text-muted-foreground">
-          {translate('auto.components.UpdateCard.6e45bfa2e0', 'Downloading...')} {percent}%
-        </p>
-      </div>
-    </div>
-  )
-}
-
-// ── Ready to install content ─────────────────────────────────────────
-
-function ReadyToInstallContent({
-  version,
-  onRestart,
-  onClose
-}: {
-  version: string
-  onRestart: () => void
-  onClose: () => void
-}) {
-  return (
-    <div className="flex flex-col gap-3 p-4">
-      <div className="flex items-start justify-between gap-2">
-        <h3 className="text-sm font-semibold">
-          {translate('auto.components.UpdateCard.17412483da', 'Ready to Install')}
-        </h3>
-        <Button
-          variant="ghost"
-          size="icon"
-          className="size-7 shrink-0 min-w-[44px] min-h-[44px] -m-2"
-          onClick={onClose}
-          aria-label={translate('auto.components.UpdateCard.8acbdd3961', 'Minimize to status bar')}
-        >
-          <Minus className="size-3.5" />
-        </Button>
-      </div>
-
-      <p className="text-sm text-muted-foreground">
-        {translate(
-          'auto.components.UpdateCard.6714206e5a',
-          "Orca v{{value0}} is downloaded. Restart when you're ready.",
-          { value0: version }
-        )}
-      </p>
-
-      <Button variant="default" size="sm" onClick={onRestart} className="w-full">
-        {translate('auto.components.UpdateCard.68b235d264', 'Restart to Update')}
-      </Button>
     </div>
   )
 }
