+++ src/renderer/src/lib/i18n-jsx-spacing-guard.test.ts
@@ -34,17 +34,17 @@ const CASES: GuardCase[] = [
     label: 'Found a setup command in source'
   },
   {
-    file: 'components/settings/RepositoryHooksSection.tsx',
+    file: 'components/settings/RepositoryHooksYamlStatus.tsx',
     afterFallback: 'Example',
     label: 'Example orca.yaml template'
   },
   {
-    file: 'components/settings/RepositoryHooksSection.tsx',
+    file: 'components/settings/RepositoryHookScriptSetting.tsx',
     afterFallback: 'Edit',
     label: 'Edit orca.yaml'
   },
   {
-    file: 'components/settings/RepositoryHooksSection.tsx',
+    file: 'components/settings/RepositoryHookPolicySettings.tsx',
     afterFallback: 'When both',
     label: 'When both orca.yaml'
   },
